requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic[email]
python-jose[cryptography]
passlib[bcrypt]
python-multipart
bcrypt==3.2.2
